third_party/blink/web_tests/virtual/controls-refresh/calendar-picker/date-picker-hotkeys.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="date" id="date-0" value="2019-02-14">
<input type="date" id="date-1" value="2019-02-14">
<input type="date" id="date-2" value="2019-02-14">
<input type="date" id="date-3" value="2019-02-14">
<script>

promise_test(() => {
  let dateElement = document.getElementById('date-0');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('t');

    let splitDate = dateElement.value.split('-');
    let actualTodayDateString = new Date(splitDate[0], splitDate[1] - 1, [splitDate[2]]).toDateString();
    let expectedTodayDateString = new Date().toDateString();

    assert_equals(actualTodayDateString, expectedTodayDateString, "Today hotkey should have updated in-page control to today's date");
    assert_equals(internals.pagePopupWindow, null, "'t' key should close popup.");
  });
}, "Date picker: 't' key should select today's date and close popup");

promise_test(() => {
  let dateElement = document.getElementById('date-1');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('m');
    assert_equals(dateElement.value, "2019-03-14", "Next month hotkey should have incremented month");

    eventSender.keyDown('M');
    assert_equals(dateElement.value, "2019-02-14", "Previous month hotkey should have decremented month");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Date picker: Previous/next month hotkey (m/M)");

promise_test(() => {
  let dateElement = document.getElementById('date-1');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('PageDown');
    assert_equals(dateElement.value, "2019-03-14", "Next month hotkey should have incremented month");

    eventSender.keyDown('PageUp');
    assert_equals(dateElement.value, "2019-02-14", "Previous month hotkey should have decremented month");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Date picker: Previous/next month hotkey (PageDown/PageUp)");

promise_test(() => {
  let dateElement = document.getElementById('date-2');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('y');
    assert_equals(dateElement.value, "2020-02-14", "Next year hotkey should have incremented year");

    eventSender.keyDown('Y');
    assert_equals(dateElement.value, "2019-02-14", "Previous year hotkey should have decremented year");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Date picker: Previous/next year hotkey");

promise_test(() => {
  let dateElement = document.getElementById('date-3');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('d');
    assert_equals(dateElement.value, "2029-02-14", "Next year hotkey should have incremented decade");

    eventSender.keyDown('D');
    assert_equals(dateElement.value, "2019-02-14", "Previous year hotkey should have decremented decade");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Date picker: Previous/next decade hotkey");

</script>
</body>
</html>
